requirements.txt
streamlit
langchain-aws
psutil
PyPDF2
langchain
langchain-experimental
langchain-community
datasets
python-docx
faiss-cpu
python-dotenv
nltk
tabulate
boto3
botocore
pandas
pytest
